+++ packages/esm-patient-queues-app/src/active-visits/active-visits-patients-reception/active-visits-reception-table.component.tsx
@@ -231,17 +231,6 @@ function ActiveVisitsReceptionTable() {
           </TableContainer>
         )}
       </DataTable>
-
-      {/* {showOverlay && (
-        <PatientSearch
-          view={view}
-          closePanel={() => setShowOverlay(false)}
-          viewState={{
-            selectedPatientUuid: viewState.selectedPatientUuid,
-          }}
-          headerTitle={overlayHeader}
-        />
-      )} */}
     </div>
   );
 }
