+++ packages/esm-patient-queues-app/src/active-visits/change-status-move-to-next-dialog.component.tsx
@@ -28,8 +28,7 @@ import { addQueueEntry, getCareProvider, updateQueueEntry } from './active-visit
 import { useQueueRoomLocations } from '../hooks/useQueueRooms';
 import styles from './change-status-dialog.scss';
 import { QueueStatus, extractErrorMessagesFromResponse } from '../utils/utils';
-import { getCurrentPatientQueueByPatientUuid, useProviders } from './visit-form/queue.resource';
-import { first } from 'rxjs/operators';
+import { getCurrentPatientQueueByPatientUuid, useProviders } from './patient-queues.resource';
 
 interface ChangeStatusDialogProps {
   patientUuid: string;
@@ -145,8 +144,7 @@ const ChangeStatusMoveToNext: React.FC<ChangeStatusDialogProps> = ({ patientUuid
 
     const abortController = new AbortController();
     updateVisit(activeVisit.uuid, endVisitPayload, abortController)
-      .pipe(first())
-      .subscribe(
+      .then(
         (response) => {
           mutate();
